Improvements

+++ app/src/main/java/com/example/t_bank/data/remote/datasource/BudgetRemoteDataSource.kt
@@ -5,6 +5,6 @@ import com.example.t_bank.data.model.BudgetStatus
 
 interface BudgetRemoteDataSource {
     suspend fun setBudget(budgetRequest: BudgetRequest)
-    suspend fun getBudgetStatus(userId: Int): BudgetStatus?
+    suspend fun getBudgetStatus(userId: Long): BudgetStatus?
     suspend fun updateIncome(userId: Int, income: Double)
 }
